fix proto name

+++ includes/miniRT.h
@@ -6,7 +6,7 @@
 /*   By: hnogared <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 18:09:56 by hnogared          #+#    #+#             */
-/*   Updated: 2023/11/23 15:07:58 by motoko           ###   ########.fr       */
+/*   Updated: 2023/11/23 15:10:52 by motoko           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ void			free_str_tab(char **str_tab);
 /* SRCS/OBJECT_MANAGEMENT */
 /* object_creation.c */
 t_object		*new_camera(t_object *to_set, t_coords coords, int fov);
-t_object		*new_light(t_object *to_set, t_coords coords, int brightness);
+t_object		*new_light(t_object *to_set, t_coords coords, float brightness);
 t_object		*new_sphere(t_object *to_set, t_coords coords, float diameter);
 t_object		*new_plane(t_object *to_set, t_coords coords);
 t_object		*new_cylinder(t_object *to_set, t_coords coords, float diameter,
